[YSA-31] Feature: 오른쪽 꼬리 말풍선 구현
